+++ CSC8503CoreClasses/CMakeLists.txt
@@ -127,7 +127,6 @@ source_group("Physics" FILES ${Physics})
 set(Header_Files
     "Debug.h"
     "GameObject.h"
-    "PlayerObject.h"
     "GameWorld.h"
     "RenderObject.h"
     "Transform.h"
@@ -139,7 +138,6 @@ source_group("Header Files" FILES ${Header_Files})
 set(Source_Files
     "Debug.cpp"
     "GameObject.cpp"
-    "PlayerObject.cpp"
     "GameWorld.cpp"
     "RenderObject.cpp"
     "Transform.cpp"
